chore: move into testDataBuilder file and exclude from test coverage
